Include some skipped jobs in the code-review report
In the recent change to omit skipped jobs when reporting, we may
have swung the pendulum too far.  While it seems that users may
not want to see a list of hundreds of skipped child_jobs, they may
want to see a list of a small number of skipped jobs due to failed
dependencies.

To try to thread the needle, we omit skipped jobs from the report
iff they were skipped due to zuul_return child_jobs; otherwise
we include them.

Change-Id: I66a223da344a93b4691a969876e887b5eec0e67c