Try zuul web builds with swc
create-react-app (CRA) is/was used to manage our zuul react web app.
CRA is an opinionated tool intended to be used with minimal
configuration. Unfortunately, it chooses to use the Babel transpiler
which is quite slow. This is a known issue [0] and others have requested
support for swc and or esbuild. Thankfully people that recognize the
utility of CRA also recognize being able to do some customization is
useful and built Create React App Config Override (CRACO) to do this.
There are CRACO plugins for both swc and esbuild. We try swc to start.

Note that we could eject the CRA app and do all this config by hand, but
considering this is largely done by CRACO we start here.

Change-Id: I8df72c4ee7a847e8f8520347f0c800e610cf4102