Add "draft" github pipeline requirement
This adds the "draft" PR status as a pipeline requirement to the
GitHub driver.  It is already used implicitly in dependent pipelines,
but this will allow it to be added explicitly to other pipelines
(for example, check).

This also fixes some minor copy/pasta errors in debug messages related
to github pipeline requirements.

Change-Id: I05f8f61aee251af24c1479274904b429baedb29d