Fix shutdown race in timer driver
If a reconfiguration happens after the timer connection is stopped,
we should not restart the apscheduler; that can lead to zombie
threads.

We should not perferm a reconfiguration after stopping the connection
anyway, so to correct that, change the shutdown sequence so that we
stop the layout update thread before we stop the connections.

Change-Id: If9c387dfc42fcba7de3969d81cf2c7917d7e4fa3