Add JobData refresh test
We try to avoid refreshing JobData from ZK when it is not necessary
(because these objects rarely change).  However, a bug in the avoidance
was recently discovered and in fact we have been refreshing them more
than necessary.

This adds a test to catch that case, along with fixing an identical
bug (the same process is used in FrozenJobs and Builds).

The fallout from these bugs may not be exceptionally large, however,
since we generally avoid refreshing FrozenJobs once a build has
started, and avoid refreshing Builds once they have completed,
meaning these bugs may have had little opportunity to show themselves.

Change-Id: I41c3451cf2b59ec18a20f49c6daf716de7f6542e