Fix pipeline event starvation due to cleanup locks
The test_web_labels_allowed_list test has been failing occasionally
due to a timeout waiting to settle.  The test ends with an unprocessed
PipelinePostConfig event in the tenant-one.check pipeline management
queue.

In these cases, the scheduler skips processing the tenant-one.check
pipeline because it is locked.  We can't see exactly why it's locked,
but based on log messages, it seems likely that one of the periodic
cleanup routines may be responsible.  These should finish quickly,
which means the starvation is likely due to a race.  I suspect the
sequence is (numbers indicate threads):

1] PipelinePostConfig event emitted
2] Cleanup routine (eg node cleanup) locks the pipeline
1] Main loop wakes up due to the event notification
1] Main loop fails to lock pipeline
1] Main loop sleeps
2] Cleanup routine completes and unlocks pipeline

At this point, the system is idle, and no more events are emitted, so
the main loop never wakes up again.

To address this, we could set the wake flag each time we encounter a
locked pipeline (we do something similiar in other cases), but on a busy
system with two schedulers, that would probably just always set the flag.
The cases where we do that also have other constraints that make the
infinite-wakeup situation less likely.

Instead, whenever we encounter a locked pipeline, let's check to see if
there are any events for that pipeline, and if so, set the wake flag
so that we run through the loop again.  This may (probably will) add
an extra loop iteration in many cases (where a second scheduler will
essentially double check that the first scheduler did finish processing
all the events while the first scheduler had the pipeline locked).  But
that shouldn't be a huge cost, and it will avoid the starvation
seen in the test, which could really happen on a nearly-idle production
system.

Change-Id: I05536389377b7753fb2430d052d188c7f2a37bff