Update patch set 3
Patch Set 3:

Thanks for the explanation.  Here is how I would recommend you proceed:

First, fix the rebase locally.  Then, if it's still substantially similar to the original, then just update the original change with the new patchset.  That will maintain the authorship and review history.  If a completely new approach is needed, then make a new change (or keep using this change) and mention the older patch in the commit message if it's relevant.

WIP changes are fine (we all upload unfinished changes in order to share our work with others).  But they still only make sense for changes that we hope to eventually review and merge, and it wasn't originally clear what the intended path was for this change.

Patch-set: 3