Add Python 3.11 testing
This adds Python 3.11 testing to Zuul. Typically we would drop python
3.10 testing at this point and only keep testing for our oldest and
newest supported versions. We don't do that here because we want to have
comparative data for performance improvements between 3.8, 3.10, and
3.11. Additionally, the current python 3.11 package available on Jammy
an RC version of 3.11. We probably want to hold off on relying on this
version of python until the package is updated to a proper release.

Change-Id: Ic5ecf2e23b250d3dbf592983b17ec715d6e9722e