wip: Initialize pipelines during pipeline processing
Calling the post-config method for a pipeline could so far be a very
expensive operation. The reason being that the post-config step was
serialized and for each pipeline the whole state was restored as a
side-effect of the init.

To speed up reconfigurations we move this initialization step from the
config loader to the actual pipeline processing. This means we can
potentially perform this step in parallel and we can avoid restoring the
pipeline state during reconfigurations as the state is not used there.

Change-Id: I4507e4d14d61b6f72c289aecf8faf9eb617e1306