Don't refresh the pipeline state during reconfig
When resetting the pipeline state after a reconfig, we will no longer
refresh the whole pipeline structure and just update the necessary
fields of the top-level pipeline state.

This can speed up reconfigurations significantly for tenants that have a
lot of active items especially with large jobs graphs.

We also add a log message right before we load the tenant after we've
aquired the tenant write lock.

This change also fixes a small bug where the znode version is
initialized to None instead of -1 when no zstat exists.

Change-Id: I7aa38569cc2f8a28758a108d4b4d8b86967da985