[WIP] Add soft version of fail-fast
fail-fast cancel all running jobs after a job has failed, and
this gives feedback fast to the user waiting for it, and safe
resources by avoiding running jobs in a change that will have
to run the jobs again. Nevertheless, there could be a benefit
from leaving jobs that already started to run finish, since
they are already using resources that could be considered wasted
if the job is cancelled, and for a limite extra time we will get
more complete feedback.

This can be considered as a middle way between fail-fast
true and false, or a less aggressive version of the fail-fast mode.

Change-Id: If39145e69a392a15f678f8c8f1d233a9c0f7eb76