Initialize tracing module in model tests
The model tests exercise some tracing code (by creating buildsets,
etc) but do not have the tracing module initialized because they are
not full Zuul system tests and do not inherit from ZuulTestCase.

Without initialization, we get different behavior from opentelemetry
(we have not provided our global tracer object, etc), which makes
these tests perform differently than they would in production.  They
also fail when run in isolation because of these differences.  They
likely only succeed when we run the whole test suite because of leakage
of the opentelemetry module initialization.

To correct this, initialize the tracing library (but with no
configuration, so we still are not configuring it to emit traces).

Change-Id: Iea5a8890f545c4e4ddf0a796b3a2fcfc7078a50e