Fix skipped builds filter in web ui
The change to automatically filter out skipped builds inadvertently
affected the buildsets tab as well since they share the query builder.
This caused the buildsets tab to automatically append a filter which
is invalid for buildsets and therefore return a 404.

This change adds an argument to the query builder to indicate whether
it should include the automatic filter.

Additionally, the query builder had an error which caused it to
append the query in all cases (which affects the otherwise working
builds tab).  This change corrects that as well.

Change-Id: Id03eee51b3551c4b32ae3c6b784b0993c4a3779c