Fix config-errors dedicated page
There was a type update to our configErrors state from a list of errors
to a dict containing a list of errors and a ready state flag. This broke
the map() call on configErrors in the dedicated config errors page
because you can't map an object only an Array. We fix this by mapping
the objects errors attribute into the page as the config errors allowing
us to keep mapping over it.

Change-Id: Ib84ca9241fb456f9493a07d1c7e1220f5fa925c8