Handle missing min_ltimes projects
If a project is missing from min_ltimes, we'll fail to update the data
correctly and will abort instead will parsing tenant details.  We can
mitigate against this by setting a default value when applying the
branch name.

Superficially, this only appears to happen when loading the state from
ZooKeeper.  All other uses of min_ltimes appear to inject a
defaultdict() instead.  This can be resolved by wrapping the result
object in a defaultdict() as well.

Change-Id: I513beab7bc6dbb1e32b4e85bc297b497c8bc7ae6