Fix race in merger shutdown
We can disconnect from ZK while the merger is still running which
can have some adverse effects and cause tests to never exit.

This moves the zk disconnect in the merger to the join method so
that we ensure that we have exited the main loop.

It also adds some improved logging so that not everything just
says "Stopped".

Change-Id: I459af85ac70ecf1f61645466d0eddc63c7e61ff9