Update patch set 6
Patch Set 6:

To address the problem of deadlocks (because we submit jobs within a running job), I elected to use multiple ThreadPoolExecutors.  In practice, with this change, we'll only use up to 2 (one for ChangeQueues, one for BuildSets), but we might add more later.  Each one by default will have max_workers set based on the number of processors, so we're probably really spawning too many threads.  We could halve the max workers, but then we would be using too few threads in the case where one of the executors is idle while the other runs.

I think observing the behavior here is worth a shot.

An alternative I considered is to use a single TPE with a priority queue.  That alone isn't enough to avoid a deadlock, but I think it could be done if we take care to make sure that we leave space when submitting jobs.  But doing all this would require an invasive subclass and I'm not very comfortable with that on a TPE.  Worth considering, but I don't think it's better than the multiple TPE approach.

Finally, we could look at using asyncio.  Structurally, it's a good fit here, but I'm not sure we should open that can of worms, and I think the tracebacks in case of an error would be especially bad here.

Patch-set: 6