docs: remove orphaed howto pages
The zuul-from-scratch page was removed with
I3c6327c9bc1a924f076ded06afc0afc4e3024384, but all these files it
linked to were left behind.

At first glance this seemed a bit odd, because sphinx should warn when
pages aren't linked to from a TOC.  It took me a while to realise
these pages were already marked with :orphan: at the top which stopped
this happening.  So they really are orhpans now, but we haven't
noticed.

This appears to go back well before the zuul-from-scratch removal to
some of the original organisation several years ago in
I206a2acf09eb8a2871ec61a00226654c798bb3eb -- it's not clear if this
flag was intended to be left in the files or was a temporary step; but
it seems that as we've gone on we've copied it into all the other
files as they got created too.

Most of this is all old and part of the bit-rot as described in the
original zuul-from-scratch removal.  The only recent part is some
console streaming docs added with
I5bfb61323bf3219168d4d014cbb9703eed230e71 -- upon re-evaluating this
I've moved it into the executor docs where it seems to fit.

The other orphaned files are removed.

Change-Id: Id3669418189f1083a2fb690ada0b60043a77b1d6