Smooth out upgrades to playbook semaphores
If running a scheduler on code prior to change
I90f5e0f570ef6c4b0986b0143318a78ddc27bbde when upgrading an executor
to a version which includes it, playbooks will lack a semaphores
list. Assume it's empty in that case, in order to avoid causing
transient build failures during a rolling upgrade of Zuul services.

Change-Id: I17d08f40e9feba62923c6f12a1d4f8b2a05f9cae