Abort job if cleanup playbook timed out
We've investigated an issue where a job was stuck on the executor
because it wasn't aborted properly. The job was cancelled by the
scheduler, but the cleanup playbook on the executor ran into a timeout.
This caused another abort via the WatchDog.

The problem is that the abort function doesn't do anything if the
cleanup playbook is running [1]. Most probably this covers the case
that we don't want to abort the cleanup playbook after a normal job
cancellation.

However, this doesn't differentiate if the abort was caused by the run
of the cleanup playbook itself, resulting in a build that's hanging
indefinitely on the executor.

To fix this, we now differentiate if the abort was caused by a stop()
call [2] or if it was caused by a timeout. In case of a timeout, we kill
the running process.

Add a test case to show the changed behaviour. Without this change, Zuul
doesn't abort the cleanup:

    <cleanup playbook is running>
    ...
    Ansible timeout exceeded: 5
    Abort: cleanup is in progress
    <cleanup playbook continues>
    ...

But with the change, it does:

    <cleanup playbook is running>
    ...
    Ansible timeout exceeded: 5
    Abort: sending kill signal to job process group
    Ansible output terminated
    <cleanup playbook stopped>

[1]: https://opendev.org/zuul/zuul/src/commit/4d555ca675d204b1d668a63fab2942a70f159143/zuul/executor/server.py#L2688
[2]: https://opendev.org/zuul/zuul/src/commit/4d555ca675d204b1d668a63fab2942a70f159143/zuul/executor/server.py#L1064

Change-Id: I979f55b52da3b7a237ac826dfa8f3007e8679932