Avoid replacing timer apscheduler jobs
If a timer trigger is configured with a large jitter and a
reconfiguration happens within the jitter time, it is possible
to miss an expected scheduled trigger because the act of
reconfiguration removes and re-adds all of a tenant's timer
trigger apscheduler jobs.

To avoid this situation, we will try to preserve any jobs with
identical configurations.

Change-Id: I5d3a4d7be891fcb4b9a3f268ee347f2069aaded3