Ignore commit message in change matchers
This change fixes an edge case that you will only encounter with Gerrit
changes that don't touch any files.

For Gerrit changes the commit message will be part of the changed files
list using the special `COMMIT_MSG` name. However, we never want a file
filter to match if the only "changed file" is the commit message.

Change-Id: I579dd6b50cd50a78d5e846f7c2376ffc9e7ba4a1