Avoid acquiring pipeline locks in manager postConfig
After creating a new tenant layout, we call _postConfig on every
pipeline manager.  That creates the shared change queues and then
also attaches new PipelineState and PipelineChangeList objects
to the new Pipeline objects on the new layout.

If the routine detects that it is the first scheduler to deal with
this pipeline under the new layout UUID, it also resets the
pipeline state in ZK (resets flags and moves all queues to the
old_queues attribute so they will be re-enqueued).  It also drops
a PostConfigEvent into the pipeline queues to trigger a run of
the pipeline after the re-enqueues.

The work in the above paragraph means that we must hold the
pipeline lock for each pipeline in turn during the reconfiguration.
Most pipelines should not be processed at this point since we do
hold the tenant write lock, however, some cleanup routines can
be operating, and so we would end up waiting for them to complete
before completing the reconfiguration.  This could end up adding
minutes to a reconfiguration.  Incidentally, these cleanup routines
are written with the expectation that they may be running during
a reconfiguration and handle missing data from refreshes.

We can avoid this by moving the "reset" work into the PipelineState
deserialization method, where we can determine at the moment we
refresh the object whether we need to "reset" it and do so.  We
can tell that a reset needs to happen if the layout uuid of the
state object does not match the running layout of the tenant.

We still need to attach new state and change list objects to the
pipeline in _postConfig (since our pipeline object is new).  We
also should make sure that the objects exist in ZK before we leave
that method, so that if a new pipeline is created, other schedulers
will be able to load the (potentially still empty) objects from ZK.
As an alternative, we could avoid even this work in _postConfig, but
then we would have to handle missing objects on refresh, and it
would not be possible to tell if the object was missing due to it
being new or due to an error.  To avoid masking errors, we keep
the current expectation that we will create these objects in ZK
on the initial reconfiguration.

Finally, we do still want to run the pipeline processing after
a reconfiguration (at least for now -- we may be approaching a
point where that will no longer be necessary).  So we move the
emission of the PostConfigEvent into the scheduler in the cases
where we know it has just updated the tenant layout.

Change-Id: Ib1e467b5adb907f93bab0de61da84d2efc22e2a7