Don't install ansible for tox linters and bindep
Previously we checked for requirements.txt in the installation lines and
assumed if it wasn't present that we are installing zuul and should run
the zuul command to install ansible as well. In reality a few other tox
targets like linters and bindep don't use requirements.txt and are not
installing zuul. We need to check for those to avoid installing ansible
unnecessarily.

Change-Id: If1d11e429a549ca7e2a4a2e195395305b8dd597a