Store paused events on the build and report them
TODO:
  - Add a release note for this change.
  - Adapt SQL reporter
  - Validate that serialization to/from ZK works (multi-scheduler)

When a build is paused, we now store this information on the build
together with its timestamp. Therefore the build now provides an "event
list" which allows us to store other events (e.g. resume or early
failures) in the future.

In addition those events are also stored in the SQL database and added
to the MQTT payload, so the information can be used by the zuul-web UI
(e.g. to show those events also in the "build times" gantt chart) or
external services.

Change-Id: I789b4f69faf96e3b8fd090a2e389df3bb9efd602