Reuse queue items after reconfiguration
When we reconfigure, we create new Pipeline objects, empty the values
in the PipelineState and then reload all the objects from ZK.  We then
re-enqueue all the QueueItems to adjust and correct the object
pointers between them (item_ahead and items_behind).  We can avoid
reloading all the objects from ZK if we keep queue items from the
previous layout and rely on the re-enqueue method correctly resetting
any relevant object pointers.

We already defer this re-enqueue work to the next pipeline processing
after a reconfiguration (so the reconfiguration itself doesn't take
very long, but now the first pipeline run after a reconfiguration must
perform a complete refresh).  With this change, then that first refresh
is no longer be a complete refresh but a normal refresh, so we will get
the benefits of previous reduction in refresh times.

The main risk of this change is that it could introduce a memory leak.
During development, additional debugging was performed to verify that
after a re-enqueue, there are no obsolete layout or pipeline objects
reachable from the pipeline state object.

For any given scheduler, there will be a period where outdated layout
or pipeline objects will be in memory: between the time that a
reconfiguration is performed and the next pipeline refresh for each
pipeline in that tenant.  This may increase memory use somewhat, but
in a bounded way (the worst case scenario is the number of obsolete
layouts in memory is equal to the number of pipelines in the system).

Change-Id: I7fa99cd83a857216321f8d946fd42abd9ec427a3